Added forceAuthn query parameter
Added support for a forceAuthn query parameter. When the value is either
'1' or 'true' then the SAML authentication request mocked up for the SP
includes the ForceAuthn option set to "true". If the SAML backend has
mirror_corce_authn configured then the ForceAuthn option will propagate
to the authenticating IdP.